+++ Cargo.toml
@@ -6,6 +6,7 @@ edition = "2024"
 [dependencies]
 anyhow = "1.0.98"
 clap = "4.5.40"
+crossterm = "0.29.0"
 id3 = "1.16.3"
 ratatui = "0.29.0"
 realfft = "3.5.0"
