+++ src/lib.rs
@@ -2,5 +2,6 @@
 
 pub mod audio;
 pub mod error;
+pub mod app;
 // pub mod Visualizer;
 // pub mod cli;
